requirements.txt
accelerate
aioboto3
aiohttp

# database packages
graspologic
hnswlib
nano-vectordb
neo4j
networkx
ollama
openai
oracledb
pymilvus
pymongo
pymysql
pyvis
# lmdeploy[all]
sqlalchemy
tenacity


# LLM packages
tiktoken
torch
transformers
xxhash

python-dotenv
google-generativeai
